+++ index.hash.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.hash.js"></script>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
