--- src/main.hash.js
@@ -1,1 +1,0 @@
-import "./main.js";
